pyproject.toml
[tool.poetry]
name = "cappuccino-discord"
version = "0.5.0"
description = ""
authors = ["Lee Watson <rev@revthefox.co.uk>"]

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^1.4.0"
sqlalchemy = { version = "^1.3.0", extras=["postgresql"] }
alembic = "^1.4.0"
pyyaml = "^5.3"
aiohttp = { version = "^3.6.2", extras=["speedups"] }
sentry-sdk = "^0.17.0"
dotty-dict = "^1.2.1"
redis = {extras = ["hiredis"], version = "^3.5.3"}

[tool.poetry.dev-dependencies]
pre-commit = "^2.7.1"
flake8 = "^3.8.3"
black = "^20.8b1"

[tool.black]
line-length = 88
target-version = ["py38"]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
